requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
httpx==0.23.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
openai==0.20.0
pandas==1.4.3
pyOpenSSL==22.0.0
requests==2.28.1
urllib3==1.26.9
transformers==4.36
Werkzeug==2.1.2
yarl==1.7.2
haystack-ai
pymed
langchain_openai
gpt_index
langchain==0.0.340
langchain-community
scipy==1.11.4
bitsandbytes==0.41.3
python-dotenv
requests